third_party/WebKit/LayoutTests/platform/win/accessibility/name-calc-aria-owns-expected.txt

This is a testharness.js-based test.
PASS Aria-owns is considered in a name computation 
Harness: the test ran to completion.
